feat: add f32 feature
- Add `f32` feature
- f64 as default precision
- Use `Float` type for `f32` and `f64`

+++ bonsai/src/bt.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 
-use crate::{state::State, ActionArgs, Behavior, Status, UpdateEvent};
+use crate::{state::State, ActionArgs, Behavior, Float, Status, UpdateEvent};
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -45,15 +45,15 @@ impl<A: Clone, B> BT<A, B> {
     /// Passes event, delta time in seconds, action and state to closure.
     /// The closure should return a status and remaining delta time.
     ///
-    /// return: (Status, f64)
+    /// return: (Status, Float)
     /// function returns the result of the tree traversal, and how long
     /// it actually took to complete the traversal and propagate the
     /// results back up to the root node
     #[inline]
-    pub fn tick<E, F>(&mut self, e: &E, f: &mut F) -> Option<(Status, f64)>
+    pub fn tick<E, F>(&mut self, e: &E, f: &mut F) -> Option<(Status, Float)>
     where
         E: UpdateEvent,
-        F: FnMut(ActionArgs<E, A>, &mut B) -> (Status, f64),
+        F: FnMut(ActionArgs<E, A>, &mut B) -> (Status, Float),
     {
         if self.finished {
             return None;
